feat: extract editor to component
