+++ playgrounds/app/src/db/schema.ts
@@ -1,4 +1,4 @@
-import { int, sqliteTable, text } from 'drizzle-orm/sqlite-core'
+import { int, sqliteTable, text, real } from 'drizzle-orm/sqlite-core'
 
 export const usersTable = sqliteTable('users_table', {
   id: text().primaryKey(),
@@ -20,6 +20,18 @@ export const snippetsTable = sqliteTable('snippets_table', {
   codeRight: text().notNull(),
   createdAt: int().notNull().default(0),
   updatedAt: int().notNull().default(0),
+  snippetWidth: int().notNull().default(450),
+  yPadding: int().notNull().default(42),
+  xPadding: int().notNull().default(42),
+  // sqlite doesn't support booleans
+  shadowEnabled: int().notNull().default(1),
+  shadowOffsetY: int().notNull().default(10),
+  shadowBlur: int().notNull().default(10),
+  shadowColor: text().notNull().default('#000000'),
+  shadowOpacity: real().notNull().default(0.6),
+  bgColor: text().notNull().default('#ffffff'),
+  language: text().notNull().default('tsx'),
+  theme: text().notNull().default('nord'),
 })
 
 export const schema = {
